# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Regatta:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      location:
        type: string
      start_date:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  model.Sailor:
    properties:
      birth_date:
        type: string
      boat_class:
        type: string
      created_at:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      regatta_id:
        type: string
      sail_number:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Backend API for the sailing competition judging application. Handles
    regattas, race groups, individual race measurements, sailors, documents, and user
    authentication.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/user/judging-of-sailing
  title: Judging of Sailing API
  version: 1.0.0
paths:
...
  /regattas:
    get:
      description: Returns all regattas ordered by start date descending.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Regatta'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List regattas
      tags:
      - regattas
    post:
      consumes:
      - application/json
      description: Creates a new regatta with title, dates and optional location.
      parameters:
      - description: Regatta details (id, created_at, updated_at are ignored)
        in: body
        name: regatta
        required: true
        schema:
          $ref: '#/definitions/model.Regatta'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Regatta'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create regatta
      tags:
      - regattas
  /regattas/{id}:
    delete:
      description: Deletes a regatta and all associated data (race groups, races,
        sailors, documents).
      parameters:
      - description: Regatta UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete regatta
      tags:
      - regattas
    get:
      description: Returns a single regatta by its UUID.
      parameters:
      - description: Regatta UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Regatta'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: regatta not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get regatta by ID
      tags:
      - regattas
    put:
      consumes:
      - application/json
      description: Updates an existing regatta. The ID in the path overrides any ID
        provided in the body.
      parameters:
      - description: Regatta UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated regatta fields (id from path is used)
        in: body
        name: regatta
        required: true
        schema:
          $ref: '#/definitions/model.Regatta'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Regatta'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update regatta
      tags:
      - regattas
...
  /regattas/{regatta_id}/sailors:
    get:
      description: Returns all sailors registered in a given regatta.
      parameters:
      - description: Regatta UUID
        in: path
        name: regatta_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Sailor'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List sailors by regatta
      tags:
      - sailors
  /regattas/{regatta_id}/sailors/find:
    get:
      description: Finds a sailor by their boat class and sail number within a specific
        regatta. Used by judges to link race measurements to sailors.
      parameters:
      - description: Regatta UUID
        in: path
        name: regatta_id
        required: true
        type: string
      - description: Boat class (e.g. Optimist, Laser)
        in: query
        name: boat_class
        required: true
        type: string
      - description: Sail number on the boat
        in: query
        name: sail_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Sailor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: sailor not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Find sailor by class and sail number
      tags:
      - sailors
  /sailors:
    post:
      consumes:
      - application/json
      description: Registers a new sailor (participant) in a regatta. The combination
        (regatta_id, boat_class, sail_number) must be unique.
      parameters:
      - description: Sailor details
        in: body
        name: sailor
        required: true
        schema:
          $ref: '#/definitions/model.Sailor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Sailor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create sailor
      tags:
      - sailors
  /sailors/{id}:
    delete:
      description: Deletes a sailor and all associated documents and race links.
      parameters:
      - description: Sailor UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete sailor
      tags:
      - sailors
    get:
      description: Returns a single sailor by UUID.
      parameters:
      - description: Sailor UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Sailor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: sailor not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sailor by ID
      tags:
      - sailors
    put:
      consumes:
      - application/json
      description: Updates an existing sailor's details. The ID in the path overrides
        any ID in the body.
      parameters:
      - description: Sailor UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated sailor fields
        in: body
        name: sailor
        required: true
        schema:
          $ref: '#/definitions/model.Sailor'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Sailor'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update sailor
      tags:
      - sailors
...
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from POST /api/v1/auth/login. Use format: Bearer
      <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"